requirements.txt
gpiozero==2.0
grovepi==1.0.4
matplotlib==3.8.0
numpy_stl==3.0.1
Pillow==10.1.0
PyQt5==5.15.10
RPi.GPIO
numpy
PyOpenGL